requirements.txt
alabaster==0.7.12
asgiref==3.2.7
Babel==2.8.0
blessings==1.7
bpython==0.19
certifi==2020.4.5.1
chardet==3.0.4
curtsies==0.3.1
Django==2.2.12
django-appconf==1.0.4
django-debug-toolbar==2.2
django-extensions==2.2.9
django-extra-views==0.11.0
django-filter==2.2.0
django-filters==0.2.1
django-haystack==2.8.1
django-oscar==2.0.4
django-oscar-api==2.0.2
django-oscar-api-checkout==1.0.0
django-phonenumber-field==2.0.1
django-redis==4.11.0
django-rest-framework==0.1.0
django-select2==7.2.3
django-solo==1.1.3
django-tables2==1.21.2
django-tenant-schemas==1.10.0
django-treebeard==4.3.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
docutils==0.16
factory-boy==2.12.0
Faker==4.0.3
greenlet==0.4.15
idna==2.9
imagesize==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
ordered-set==3.1.1
packaging==20.3
phonenumbers==8.12.1
Pillow==7.1.1
psycopg2-binary==2.8.5
purl==1.5
pycountry==19.8.18
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2019.3
redis==3.4.1
requests==2.23.0
six==1.14.0
snowballstemmer==2.0.0
sorl-thumbnail==12.6.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.9
wcwidth==0.1.9
